modules/locations-worker/requirements.txt
geoalchemy2==0.16.0
kafka-python==2.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
numpy==2.1.3
packaging==24.2
psycopg2-binary==2.9.10
shapely==1.7.1
six==1.16.0
sqlalchemy==2.0.36
typing-extensions==4.12.2
